src/main/webapp/app/entities/fast-food/update/fast-food-update.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of, Subject, from } from 'rxjs';

import { FastFoodService } from '../service/fast-food.service';
import { IFastFood, FastFood } from '../fast-food.model';
import { IMenu } from 'app/entities/menu/menu.model';
import { MenuService } from 'app/entities/menu/service/menu.service';

import { FastFoodUpdateComponent } from './fast-food-update.component';

describe('FastFood Management Update Component', () => {
  let comp: FastFoodUpdateComponent;
  let fixture: ComponentFixture<FastFoodUpdateComponent>;
  let activatedRoute: ActivatedRoute;
  let fastFoodService: FastFoodService;
  let menuService: MenuService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      declarations: [FastFoodUpdateComponent],
      providers: [
        FormBuilder,
        {
          provide: ActivatedRoute,
          useValue: {
            params: from([{}]),
          },
        },
      ],
    })
      .overrideTemplate(FastFoodUpdateComponent, '')
      .compileComponents();

    fixture = TestBed.createComponent(FastFoodUpdateComponent);
    activatedRoute = TestBed.inject(ActivatedRoute);
    fastFoodService = TestBed.inject(FastFoodService);
    menuService = TestBed.inject(MenuService);

    comp = fixture.componentInstance;
  });

  describe('ngOnInit', () => {
    it('Should call Menu query and add missing value', () => {
      const fastFood: IFastFood = { id: 456 };
      const menu: IMenu = { id: 18387 };
      fastFood.menu = menu;

      const menuCollection: IMenu[] = [{ id: 1855 }];
      jest.spyOn(menuService, 'query').mockReturnValue(of(new HttpResponse({ body: menuCollection })));
      const additionalMenus = [menu];
      const expectedCollection: IMenu[] = [...additionalMenus, ...menuCollection];
      jest.spyOn(menuService, 'addMenuToCollectionIfMissing').mockReturnValue(expectedCollection);

      activatedRoute.data = of({ fastFood });
      comp.ngOnInit();

      expect(menuService.query).toHaveBeenCalled();
      expect(menuService.addMenuToCollectionIfMissing).toHaveBeenCalledWith(menuCollection, ...additionalMenus);
      expect(comp.menusSharedCollection).toEqual(expectedCollection);
    });

    it('Should update editForm', () => {
      const fastFood: IFastFood = { id: 456 };
      const menu: IMenu = { id: 19147 };
      fastFood.menu = menu;

      activatedRoute.data = of({ fastFood });
      comp.ngOnInit();

      expect(comp.editForm.value).toEqual(expect.objectContaining(fastFood));
      expect(comp.menusSharedCollection).toContain(menu);
    });
  });

  describe('save', () => {
    it('Should call update service on save for existing entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<FastFood>>();
      const fastFood = { id: 123 };
      jest.spyOn(fastFoodService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ fastFood });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: fastFood }));
      saveSubject.complete();

      // THEN
      expect(comp.previousState).toHaveBeenCalled();
      expect(fastFoodService.update).toHaveBeenCalledWith(fastFood);
      expect(comp.isSaving).toEqual(false);
    });

    it('Should call create service on save for new entity', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<FastFood>>();
      const fastFood = new FastFood();
      jest.spyOn(fastFoodService, 'create').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ fastFood });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.next(new HttpResponse({ body: fastFood }));
      saveSubject.complete();

      // THEN
      expect(fastFoodService.create).toHaveBeenCalledWith(fastFood);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).toHaveBeenCalled();
    });

    it('Should set isSaving to false on error', () => {
      // GIVEN
      const saveSubject = new Subject<HttpResponse<FastFood>>();
      const fastFood = { id: 123 };
      jest.spyOn(fastFoodService, 'update').mockReturnValue(saveSubject);
      jest.spyOn(comp, 'previousState');
      activatedRoute.data = of({ fastFood });
      comp.ngOnInit();

      // WHEN
      comp.save();
      expect(comp.isSaving).toEqual(true);
      saveSubject.error('This is an error!');

      // THEN
      expect(fastFoodService.update).toHaveBeenCalledWith(fastFood);
      expect(comp.isSaving).toEqual(false);
      expect(comp.previousState).not.toHaveBeenCalled();
    });
  });

  describe('Tracking relationships identifiers', () => {
    describe('trackMenuById', () => {
      it('Should return tracked Menu primary key', () => {
        const entity = { id: 123 };
        const trackResult = comp.trackMenuById(0, entity);
        expect(trackResult).toEqual(entity.id);
      });
    });
  });
});
